refactor: remove absent prng
The absent implementation of prng was removed. Now the only fake prng
implementation will be used.

The current implementation of prng is very simple and must be
re-implemented in a production system.

The absent implementation for prng was used before for some platforms.
In this case, the branch protection and MTE features (if enabled) were
initialized with zero value. This change is needed to have a non-zero
value from the prng function in any case.

Change-Id: Ic29994e5cc3a8202760bc2f966fa4650b060f324
Signed-off-by: Dmitrii Martynov <fkr655@yandex.ru>